requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.3
cloudinary==1.26.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.0.8
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-etc==1.3.1
django-heroku==0.3.1
django-hitcount==1.3.5
django-jazzmin==2.4.9
django-js-asset==1.2.2
django-rosetta==0.9.8

gunicorn==20.1.0
idna==3.2
libretranslatepy==2.1.1
lxml==4.7.1
Pillow==8.3.1
polib==1.1.1
psycopg2==2.9.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6

